potion.h: Added 'long' cast to PN_INT().

+++ core/potion.h
@@ -111,7 +111,7 @@ struct PNVtable;
 #define PN_IS_REF(v)     (PN_TYPE(v) == PN_TWEAK)
 
 #define PN_NUM(i)       ((PN)((((long)(i))<<1) + PN_FNUMBER))
-#define PN_INT(x)       (((long)(x))>>1)
+#define PN_INT(x)       ((long)((long)(x))>>1)
 #define PN_PREC 16
 #define PN_RAND()       PN_NUM(potion_rand_int())
 #define PN_STR_PTR(x)   potion_str_ptr(x)
